PythonAndHtmlAndDB/requirements.txt
pandas
mysql.connector
flask